Some Fixes and Cave Walls,Button
Warja,

I changed the buttons' places to the bottom, because otherwise they are too close and create desire to push just for curiousity, but I might accidentally fucked up the first level's length, can you check it?

Also I changed the platform's place in 3rd level, because it was creating some vulgar scenes like buildings are beating you after some time. I tried to stay longer in the other one and looks unproblematic for now.

Caves' walls are closer to each other now, but maybe can be smaller as length since it might me klostrophobic for some people now. Also added the button for cave.

Hope you see this in the morning, good night! And then we can build,